+++ CHANGELOG.md
@@ -21,6 +21,8 @@ in 3.17. Use `PaymentSettings.defaultTransactionFlowStrategy` instead.
   - Inform apps if customer should be notified when fulfillment is created.
 - Add `NOTIFY_CUSTOMER` flag to `FulfillmentApproved` type - #13637, by @Air-t
   - Inform apps if customer should be notified when fulfillment is approved.
+- Add `GIFT_CARD_EXPORT_COMPLETED` webhook - #13765, by @Air-t
+  - Event sent when CSV export is completed.
 
 ### Other changes
 - Fix error in variant available stock calculation - 13593 by @awaisdar001